# frontend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 慧眼医疗云平台 V2.0 API
  description: |
    糖尿病视网膜病变 AI 辅助筛查 + 眼科医学影像培训考核

    **后端框架推荐**：FastAPI / Spring Boot / NestJS
    **协议**：REST + JSON · UTF-8
    **认证**：`Authorization: Bearer <token>`

    ## 统一响应结构

    所有接口（除文件下载、图片流外）统一返回：

    ```json
    {
      "code": 0,
      "msg": "success",
      "data": {}
    }
    ```

    - `code === 0` 或 `200` 表示成功
    - `code === 401` 或 `40101` 表示会话失效，前端自动跳登录
    - 其它非零 `code` 表示业务失败，前端使用 `msg` 弹错误提示

    ## 错误状态码

    | HTTP | 含义 |
    |---|---|
    | 400 | 请求参数错误 |
    | 401 | 未登录 / 会话失效 |
    | 403 | 无权限 |
    | 404 | 资源不存在 |
    | 413 | 上传文件过大 |
    | 500 | 服务器错误 |
    | 502/503/504 | 网关/服务不可用 |
  version: 2.0.0
  contact:
    name: 慧眼医疗 · 前端工程组
  license:
    name: Proprietary
servers:
  - url: http://127.0.0.1:8000
    description: 本地开发
  - url: https://huiyan.example.com/api
    description: 生产（Nginx 反向代理）
...
security:
  - bearerAuth: []
paths:
...
  /training/cases:
    get:
      tags: [Training · 医学培训]
      summary: 获取病例分页列表
      parameters:
        - { name: keyword, in: query, schema: { type: string } }
        - { name: drLevel, in: query, schema: { type: integer, enum: [0, 1, 2, 3, 4] } }
        - { name: difficulty, in: query, schema: { $ref: '#/components/schemas/Difficulty' } }
        - { name: done, in: query, schema: { type: boolean } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 100 } }
      responses:
        '200':
          description: 病例分页
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResult'
                  - properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PageResult'
                          - properties:
                              list:
                                type: array
                                items: { $ref: '#/components/schemas/TrainingCase' }
  /training/cases/{caseId}:
    get:
      tags: [Training · 医学培训]
      summary: 获取病例详情
      parameters:
        - { name: caseId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 病例详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResult'
                  - properties:
                      data:
                        $ref: '#/components/schemas/TrainingCase'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 在 `Authorization` 头中以 `Bearer <token>` 形式提交
...
  schemas:
...
    ApiResult:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
          example: 0
          description: 业务码（0/200=成功，401=未登录，其它=失败）
        msg:
          type: string
          example: success
        data:
          description: 业务数据，按各接口具体字段
          nullable: true
    PageResult:
      type: object
      required: [total, page, pageSize, list]
      properties:
        total: { type: integer, example: 128 }
        page: { type: integer, example: 1 }
        pageSize: { type: integer, example: 20 }
        list:
          type: array
          items: { type: object }

    # ===================== 枚举 =====================
...
    EyeSide:
      type: string
      enum: [OD, OS, OU]
      description: OD=右眼 / OS=左眼 / OU=双眼
    Gender:
      type: string
      enum: ['男', '女']
    Difficulty:
      type: string
      enum: ['入门', '初级', '中级', '高级']
    LesionType:
      type: string
      enum: ['出血', '渗出', '微动脉瘤', '棉绒斑', '新生血管']
...
    DRLevel:
      type: integer
      enum: [0, 1, 2, 3, 4]
      description: |
        0 = 无 DR
        1 = 轻度 NPDR
        2 = 中度 NPDR
        3 = 重度 NPDR
        4 = PDR（增殖性）

    # ===================== 认证 =====================
...
    Lesion:
      type: object
      required: [type, count]
      properties:
        type: { $ref: '#/components/schemas/LesionType' }
        count: { type: integer }
        location: { type: string, example: 黄斑区上方 }
    TrainingCase:
      type: object
      required: [id, name, age, gender, eye, drGrade, drLevel, difficulty, done, diabetesYears, hospital, lesions]
      properties:
        id: { type: string, example: CASE001 }
        name: { type: string, example: 王建国 }
        age: { type: integer }
        gender: { $ref: '#/components/schemas/Gender' }
        eye: { $ref: '#/components/schemas/EyeSide' }
        drGrade: { type: string, example: '4 级 PDR（增殖性）' }
        drLevel: { $ref: '#/components/schemas/DRLevel' }
        difficulty: { $ref: '#/components/schemas/Difficulty' }
        done: { type: boolean }
        thumbUrl: { type: string }
        imageUrl: { type: string, format: uri, description: 原图（眼底彩照） URL }
        diabetesYears: { type: integer }
        hospital: { type: string }
        bestIou:
          type: number
          format: float
          nullable: true
        lesions:
          type: array
          items: { $ref: '#/components/schemas/Lesion' }
        createdAt: { type: string }
...